ZAP_CLI/requirements.txt
zaproxy>=0.1.0
python-dotenv>=1.0.0
typer>=0.9.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-astradb>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.32.0